requirements.txt
fastapi
uvicorn
pydantic
python-jose
redis
openai
python-dotenv
pytest
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
